backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pyscf>=2.3.0
numpy>=1.23.0
scipy>=1.10.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
